Move analyzers to own folder
