+++ tests/Vogen.Tests/DiagnosticsTests/DisallowDefaultingTests.cs
@@ -2,6 +2,7 @@
 using FluentAssertions;
 using Microsoft.CodeAnalysis;
 using VerifyXunit;
+using Vogen.Analyzers;
 using Xunit;
 using Xunit.Abstractions;
